user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "User and Identity Service for Panchayat"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0",
    "redis>=5.0.3",
    "aiokafka>=0.10.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-exporter-otlp>=1.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "testcontainers[postgres]>=4.14.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]
